+++ changelog.md
@@ -24,12 +24,16 @@ Changelog -- where I'll keep progress updates.
 12/08 12:23 -- updated is_black_king_in_check and black_candidate_moves
 12/08 12:52 -- debugged black_king_in_check (it was a variable error bruh) and did send_black_move. i think random bot is ready to be coded :D 
 
+12/08 14:30 -- RANDOM MOVER BOT IS DONE!! YOU CAN NOW PLAY CHESS AGAINST IT
+12/08 14:40 -- Fixed en passant bug RRRRRRAAAAHHHHHHHHHH
+
+
 
 todolist (so i dont forget when i wake up):    
-- finish everything for black
-- make random mover bot!  
-- make a wrapper class (again) to actually play chess
+- (nothing)
 
+
+things to consider doing:
 - deploy to lichess?
 
 other also kinda important: